requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
PySocks==1.7.1
playwright==1.49.1
groq==0.9.0
pymongo[srv]==4.8.0
dnspython==2.6.1
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.4
pydantic==2.7.4
typing-extensions==4.12.2
lxml==5.2.2
aiohttp==3.9.5
